move some method of Deck back to deck.go
