+++ poolutil_test.go
@@ -8,9 +8,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/k1LoW/errors"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/api/drive/v3"
 )
 
 const (
@@ -109,22 +107,4 @@ func AcquirePresentation(t *testing.T) string {
 // ReleasePresentation returns a presentation ID to the pool.
 func ReleasePresentation(id string) {
 	presentationPool <- id
-}
-
-// AllowReadingByAnyone sets the permission of the presentation to allow anyone to read it.
-func (d *Deck) AllowReadingByAnyone(ctx context.Context) (err error) {
-	defer func() {
-		err = errors.WithStack(err)
-	}()
-	if d.id == "" {
-		return fmt.Errorf("presentation ID is not set")
-	}
-	permission := &drive.Permission{
-		Type: "anyone",
-		Role: "reader",
-	}
-	if _, err := d.driveSrv.Permissions.Create(d.id, permission).Context(ctx).Do(); err != nil {
-		return fmt.Errorf("failed to set permission: %w", err)
-	}
-	return nil
 }
